Now have unicast working from the Mac to the device. The solution is that the device tells us all three things: IP address, hostname (which may default to IP address), and the contents of /etc/hostname (which should definitely be the thing we told it to be, something like hb-<MAC>). The latter we are referring to as "deviceName", not to be confused with hostname, even through it *should* be the device's hostname. The LocalDeviceRepresentation tries various address solutions when asked to send unicast. Currently it tries hostname, followed by deviceName with .local appended. The latter is what gets through on a Mac.
TODO: we should make this more robust and efficient. Proposal: we should maintain a history of what String is the winning String in LocalDeviceRepresentation. We should always try that first, and on failure we should iterate through the various candidates again, and choose a new best String.

Also, we'll need to make the same fix-up for the code-sending stuff.

Lastly, we should check that the /device/alive message fully updates the data stored by LocalDeviceRepresentation.