compiled plugin on win10,
networking seems to have basic functionality but not displaying correctly yet. Some work before itll be in latest intelliJ. versioned and figured out adding menu items. May need to roll back on a few things but just needed to save this as a starting point to try dev the menu and panel